rating/requirements.txt
django
pika
psycopg2
djangorestframework
django-filter
python-crontab
django-redis
mock